requirements.txt
sqlalchemy
pandas
openpyxl
psycopg2-binary
ShopifyAPI
pdfplumber
better-exceptions
loguru
colorama